Extract all `Response` classes into a subpackage
