+++ src/main/java/ch/usi/si/seart/converter/StringToGraphQlErrorResponseErrorTypeConverter.java
@@ -1,6 +1,6 @@
 package ch.usi.si.seart.converter;
 
-import ch.usi.si.seart.github.GraphQlErrorResponse;
+import ch.usi.si.seart.github.response.GraphQlErrorResponse;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.lang.Nullable;
