+++ src/main/java/ch/usi/si/seart/github/response/BlockedRestErrorResponse.java
@@ -1,4 +1,4 @@
-package ch.usi.si.seart.github;
+package ch.usi.si.seart.github.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
