+++ src/main/java/ch/usi/si/seart/github/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ch.usi.si.seart.github;
+package ch.usi.si.seart.github.response;
 
 public interface ErrorResponse extends Response {
