+++ src/main/java/ch/usi/si/seart/github/response/JsonResponse.java
@@ -1,4 +1,4 @@
-package ch.usi.si.seart.github;
+package ch.usi.si.seart.github.response;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
